Delete cookies made optional

+++ seleniumbuilder/content/html/builder/record.js
@@ -136,10 +136,12 @@ builder.record.startRecording = function(urlText, seleniumVersion) {
     jQuery("#startup-url").focus();
     return;
   }
-    
-  // Delete cookies for given URL.
-  builder.deleteURLCookies(url.href());
-  
+
+  // Delete cookies for given URL if the option is checked
+  if (jQuery("#delete-cookies").prop("checked")) {
+    builder.deleteURLCookies(url.href());
+  }
+
   // Now load the page - both to ensure we're on the right page when we start recording
   // and to ensure that we get a clean page free of cookie influence.
   // Don't show the record interface until the new page has loaded.
@@ -152,7 +154,7 @@ builder.record.startRecording = function(urlText, seleniumVersion) {
     } else {
       jQuery('#heading-record').removeClass('is-on');
       if (isLoading) {
-        builder.record.recording = true;    
+        builder.record.recording = true;
         builder.gui.switchView(builder.views.script);
         builder.suite.addScript(new builder.Script(seleniumVersion));
         builder.getScript().saveRequired = true;
